backend/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
python-jose[cryptography]
httpx
pydantic
python-dotenv
yfinance
pandas
numpy
scipy
statsmodels
cachetools
